+++ backend/src/api/v1/auth/schemas.py
@@ -36,17 +36,3 @@ class UserCreate(UserEmail):
     """Represents user registration details."""
 
     password: str = Field(min_length=8, max_length=128)
-
-
-class VerifyUser(UserEmail):
-    code: int
-
-
-class NewPassword(VerifyUser):
-    """Represents new password"""
-
-    new_password: str = Field(min_length=8, max_length=128)
-
-
-class SuccessVerifyMessage(BaseModel):
-    detail: str = "The code is accepted"
