+++ backend/src/api/v1/verify/__init__.py
@@ -1,0 +1,5 @@
+from src.api.v1.verify.routes import router
+
+__all__ = [
+    "router",
+]
