+++ backend/src/api/v1/verify/schemas.py
@@ -1,0 +1,21 @@
+from pydantic import BaseModel, EmailStr, Field
+
+
+class UserEmail(BaseModel):
+    """Represents user with email."""
+
+    email: EmailStr = Field(max_length=255)
+
+
+class VerifyUser(UserEmail):
+    code: int
+
+
+class NewPassword(VerifyUser):
+    """Represents new password"""
+
+    new_password: str = Field(min_length=8, max_length=128)
+
+
+class SuccessVerifyMessage(BaseModel):
+    detail: str = "The code is accepted"
